Changed number_periods to_f to to_i

+++ week-1/day-2/compound-interest.rb
@@ -5,7 +5,7 @@ puts ("Interest rate(%):")
 interest_rate = gets.chomp.to_f
 
 puts ("Number of periods:")
-number_periods = gets.chomp.to_f
+number_periods = gets.chomp.to_i
 
 i = 0
